Fix dataKey to json conversion in surveys
